# Conflicts:
#	bench/libecrt.c
#	klippy/extras/ethercat_node.py
#	klippy/extras/servo_axis.py
#	klippy/motion_bridge.py
#	rust/kalico-ethercat-rt/src/bin/kalico-ethercat-rt-stub.rs
#	rust/kalico-ethercat-rt/src/bin/kalico-ethercat-rt.rs
#	rust/kalico-ethercat-rt/src/ffi.rs
#	rust/kalico-ethercat-rt/src/wire.rs
#	rust/kalico-ethercat-rt/src/wire/tests.rs
#	rust/kalico-ethercat-rt/tests/real_socket_streaming.rs
#	rust/kalico-ethercat-rt/tests/stub_loop.rs
#	rust/kalico-protocol/src/messages.rs
#	rust/motion-bridge/src/bridge.rs

+++ rust/kalico-c-api/include/kalico_nurbs.h
@@ -60,7 +60,7 @@ uint32_t kalico_runtime_get_dispatcher_floor_cycles(void);
 
 int32_t kalico_runtime_get_heartbeat(kalico_nurbs_KalicoRuntime *rt,
                                      uint8_t *out_engine_state,
-                                     uint8_t *out_fault_code,
+                                     uint16_t *out_fault_code,
                                      uint32_t *out_retired,
                                      uintptr_t max_axes);
